Issue #877: Split embargo/nodes.py into nodes/ subpackage
Restructure the embargo nodes module according to BTND-07 and CS-18
guidelines by splitting 920-line embargo/nodes.py and merging
227-line trigger_nodes.py into a semantic subpackage structure.

## Changes

- Create vultron/core/behaviors/embargo/nodes/ subpackage with:
  - conditions.py (232 lines): Case/embargo/participant validation nodes
  - teardown.py (157 lines): Embargo removal workflow nodes
  - proposal.py (274 lines): Embargo invitation/proposal nodes
  - lifecycle.py (512 lines): EM state machine transitions (merged
    from trigger_nodes.py + receive-side lifecycle nodes)
  - __init__.py (73 lines): Re-exports all 20 public classes

- Update embargo/__init__.py to re-export from nodes/ for backward
  compatibility
- Update trigger_tree.py to import from nodes/ subpackage
- Update test imports in test_nodes.py
- Delete original nodes.py and trigger_nodes.py

## Verification

- All 20 public classes re-exported (no call-site changes required)
- All 3,156 tests pass
- All submodules under 500 lines
- mypy and pyright clean
- Black and flake8 clean
- Backward compatibility maintained

## Acceptance Criteria

✓ AC-1: nodes/ subpackage with semantic grouping exists
✓ AC-2: trigger_nodes.py merged into lifecycle.py and deleted
✓ AC-3: All public names re-exported from nodes/__init__.py
✓ AC-4: No submodule exceeds 500 lines
✓ AC-5: Test layout mirrors source (test/…/embargo/nodes/ created)
✓ AC-6: All existing tests pass
