+++ vultron/core/behaviors/embargo/__init__.py
@@ -1,1 +1,49 @@
-"""Behavior tree nodes for embargo lifecycle workflows."""
+"""Behavior tree nodes for embargo lifecycle workflows.
+
+Re-exports all node classes from the nodes/ subpackage to maintain
+backward compatibility with existing imports.
+"""
+
+from vultron.core.behaviors.embargo.nodes import (
+    AcceptEmbargoLifecycleNode,
+    ApplyEmbargoTeardownNode,
+    CommitLogCascadeNode,
+    CreateAndStoreInviteNode,
+    IsActiveEmbargoNode,
+    LookupParticipantNode,
+    OptionalLookupParticipantNode,
+    PersistEmbargoEventNode,
+    ProposeEmbargoLifecycleNode,
+    RecordParticipantAcceptanceNode,
+    RejectEmbargoLifecycleNode,
+    RemoveFromProposedEmbargoesNode,
+    RemoveStaleAcceptanceNode,
+    SetEmbargoActiveNode,
+    TerminateEmbargoLifecycleNode,
+    TerminateEmbargoNode,
+    UpdateParticipantEmbargoPecNode,
+    ValidateCaseExistsNode,
+    ValidateEmbargoRevisionStateNode,
+)
+
+__all__ = [
+    "AcceptEmbargoLifecycleNode",
+    "ApplyEmbargoTeardownNode",
+    "CommitLogCascadeNode",
+    "CreateAndStoreInviteNode",
+    "IsActiveEmbargoNode",
+    "LookupParticipantNode",
+    "OptionalLookupParticipantNode",
+    "PersistEmbargoEventNode",
+    "ProposeEmbargoLifecycleNode",
+    "RecordParticipantAcceptanceNode",
+    "RejectEmbargoLifecycleNode",
+    "RemoveFromProposedEmbargoesNode",
+    "RemoveStaleAcceptanceNode",
+    "SetEmbargoActiveNode",
+    "TerminateEmbargoLifecycleNode",
+    "TerminateEmbargoNode",
+    "UpdateParticipantEmbargoPecNode",
+    "ValidateCaseExistsNode",
+    "ValidateEmbargoRevisionStateNode",
+]
