+++ vultron/core/behaviors/embargo/nodes/__init__.py
@@ -1,0 +1,73 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""Embargo BT nodes subpackage.
+
+Re-exports all public node classes from submodules for backward compatibility.
+"""
+
+from vultron.core.behaviors.embargo.nodes.conditions import (
+    IsActiveEmbargoNode,
+    LookupParticipantNode,
+    OptionalLookupParticipantNode,
+    ValidateCaseExistsNode,
+)
+from vultron.core.behaviors.embargo.nodes.lifecycle import (
+    AcceptEmbargoLifecycleNode,
+    CommitLogCascadeNode,
+    PersistEmbargoEventNode,
+    ProposeEmbargoLifecycleNode,
+    RejectEmbargoLifecycleNode,
+    SetEmbargoActiveNode,
+    TerminateEmbargoLifecycleNode,
+    TerminateEmbargoNode,
+    ValidateEmbargoRevisionStateNode,
+)
+from vultron.core.behaviors.embargo.nodes.proposal import (
+    CreateAndStoreInviteNode,
+    RecordParticipantAcceptanceNode,
+    RemoveStaleAcceptanceNode,
+    UpdateParticipantEmbargoPecNode,
+)
+from vultron.core.behaviors.embargo.nodes.teardown import (
+    ApplyEmbargoTeardownNode,
+    RemoveFromProposedEmbargoesNode,
+)
+
+__all__ = [
+    # Conditions
+    "ValidateCaseExistsNode",
+    "IsActiveEmbargoNode",
+    "LookupParticipantNode",
+    "OptionalLookupParticipantNode",
+    # Teardown
+    "ApplyEmbargoTeardownNode",
+    "RemoveFromProposedEmbargoesNode",
+    # Proposal
+    "UpdateParticipantEmbargoPecNode",
+    "CreateAndStoreInviteNode",
+    "RecordParticipantAcceptanceNode",
+    "RemoveStaleAcceptanceNode",
+    # Lifecycle
+    "PersistEmbargoEventNode",
+    "ValidateEmbargoRevisionStateNode",
+    "ProposeEmbargoLifecycleNode",
+    "AcceptEmbargoLifecycleNode",
+    "RejectEmbargoLifecycleNode",
+    "TerminateEmbargoLifecycleNode",
+    "TerminateEmbargoNode",
+    "CommitLogCascadeNode",
+    "SetEmbargoActiveNode",
+]
