+++ vultron/core/behaviors/embargo/nodes/teardown.py
@@ -1,0 +1,159 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""Embargo removal and teardown nodes."""
+
+import py_trees
+from py_trees.common import Status
+
+from vultron.core.behaviors.helpers import DataLayerAction
+from vultron.core.models.protocols import is_case_model
+from vultron.core.states.em import EM, is_valid_em_transition
+from vultron.core.use_cases._helpers import (
+    reset_case_participant_embargo_consent,
+)
+
+
+class ApplyEmbargoTeardownNode(DataLayerAction):
+    """Apply receiver-side embargo teardown.
+
+    Performs the ACTIVE/REVISE → EXITED EM state transition, clears
+    ``active_embargo``, and resets all participant embargo consent states.
+    Handles idempotency: if EM state is already EXITED, logs and returns
+    SUCCESS without modifying the DataLayer.
+
+    For unexpected EM states a state-sync override is applied (the sender
+    is authoritative) with a WARNING log entry, mirroring the pattern used
+    by ``AddEmbargoEventToCaseReceivedUseCase``.
+
+    When ``case_id`` is not provided at construction (``None``), the node
+    reads it from the log entry in the blackboard ``activity`` key.  This
+    allows the node to be shared between the ``RemoveEmbargoFromCaseBT``
+    (construction-time ``case_id``) and the
+    ``AnnounceLogEntryReceivedBT`` participant subtree (blackboard
+    ``activity``).
+    """
+
+    def __init__(self, case_id: str | None = None, name: str | None = None):
+        super().__init__(name=name or self.__class__.__name__)
+        self.case_id = case_id
+
+    def setup(self, **kwargs: object) -> None:
+        super().setup(**kwargs)
+        if self.case_id is None:
+            self.blackboard.register_key(
+                key="activity", access=py_trees.common.Access.READ
+            )
+
+    def update(self) -> Status:
+        if self.datalayer is None:
+            self.feedback_message = "DataLayer not available"
+            return Status.FAILURE
+
+        if self.case_id is not None:
+            case_id = self.case_id
+        else:
+            from vultron.core.behaviors.sync.nodes import _require_log_entry
+
+            entry = _require_log_entry(self.blackboard.activity, self.name)
+            case_id = entry.case_id
+
+        case = self.datalayer.read(case_id)
+        if not is_case_model(case):
+            self.feedback_message = f"Case '{case_id}' not found"
+            self.logger.warning("%s: %s", self.name, self.feedback_message)
+            return Status.SUCCESS
+
+        current_em = case.current_status.em_state
+
+        if current_em == EM.EXITED:
+            self.feedback_message = (
+                f"Case '{case_id}' EM already EXITED — idempotent no-op"
+            )
+            self.logger.info("%s: %s", self.name, self.feedback_message)
+            return Status.SUCCESS
+
+        if not is_valid_em_transition(current_em, EM.EXITED):
+            self.logger.warning(
+                "%s: EM transition %s → EXITED is not a standard machine"
+                " transition for case '%s'; applying state-sync override",
+                self.name,
+                current_em,
+                case_id,
+            )
+
+        case.current_status.em_state = EM.EXITED
+        case.active_embargo = None
+        reset_case_participant_embargo_consent(self.datalayer, case)
+        self.datalayer.save(case)
+
+        self.feedback_message = (
+            f"Embargo teardown applied on case '{case_id}'"
+            f" (EM {current_em} → EXITED)"
+        )
+        self.logger.info("%s: %s", self.name, self.feedback_message)
+        return Status.SUCCESS
+
+
+class RemoveFromProposedEmbargoesNode(DataLayerAction):
+    """Remove the embargo from the case's proposed_embargoes list.
+
+    Idempotent cleanup: returns SUCCESS if embargo successfully removed or was
+    not in proposed_embargoes (nothing to remove). Returns FAILURE only if the
+    case cannot be read (critical prerequisite missing).
+
+    Saves the case only when a change is made.
+    """
+
+    def __init__(self, case_id: str, embargo_id: str, name: str | None = None):
+        super().__init__(name=name or self.__class__.__name__)
+        self.case_id = case_id
+        self.embargo_id = embargo_id
+
+    def update(self) -> Status:
+        if self.datalayer is None:
+            self.feedback_message = "DataLayer not available"
+            return Status.FAILURE
+
+        case = self.datalayer.read(self.case_id)
+        if not is_case_model(case):
+            self.feedback_message = f"Case '{self.case_id}' not found"
+            return Status.FAILURE
+
+        def _as_id(obj: object) -> str | None:
+            return obj if isinstance(obj, str) else getattr(obj, "id_", None)
+
+        proposed_ids = [_as_id(e) for e in case.proposed_embargoes]
+        if self.embargo_id in proposed_ids:
+            case.proposed_embargoes = [
+                e
+                for e in case.proposed_embargoes
+                if _as_id(e) != self.embargo_id
+            ]
+            self.datalayer.save(case)
+            self.feedback_message = (
+                f"Removed embargo '{self.embargo_id}' from proposed_embargoes"
+                f" of case '{self.case_id}'"
+            )
+            self.logger.info(
+                "RemoveFromProposedEmbargoes: %s", self.feedback_message
+            )
+        else:
+            self.feedback_message = (
+                f"Embargo '{self.embargo_id}' not in proposed_embargoes"
+                f" of case '{self.case_id}' — nothing to remove"
+            )
+
+        return Status.SUCCESS
